Test that initializing an erroneously coerced type doesn't die
Dave Rolsky's post-1.08 changes made Moose warn on various things that
didn't warn before, but one of these warnings erroneously died. That
was fixed in "Always check that a constraint has a coercion before
calling ->coerce". This is a test to make sure it doesn't happen
again.

This affected Hailo, whose tests died on the master branch of Moose
before Dave's fix. Its behavior was fixed in Hailo commit
4dbac78b07a36 [0].

0. http://github.com/hailo/hailo/commit/4dbac78b07a36